Update Reference.md to fix typo
Under server options -> state -> cookies -> parse the docs say that the 'Cookie' headers are parsed and stored in the request.cookies object. However, I find that they are actually stored in the request.state object. Furthermore, request.state is referenced as the place to find parsed cookies later on in the file. So, I have assumed request.cookies is a typo and am submitting this as a correction.